+++ src/sharedstate.h
@@ -92,7 +92,7 @@ struct SharedState
 
 	/* Returns global quad IBO, and ensures it has indices
 	 * for at least minSize quads */
-	void ensureQuadIBO(int minSize);
+	void ensureQuadIBO(size_t minSize);
 	GlobalIBO &globalIBO();
 
 	/* Global general purpose texture */
